+++ docs/gitbook/changelog.md
@@ -1,3 +1,17 @@
+# [5.73.0](https://github.com/taskforcesh/bullmq/compare/v5.72.1...v5.73.0) (2026-04-03)
+
+
+### Features
+
+* **sandbox:** add getDependenciesCount proxy for jobs ([#3923](https://github.com/taskforcesh/bullmq/issues/3923)) ref [#3533](https://github.com/taskforcesh/bullmq/issues/3533) (elixir) (php) (python) ([cbfe437](https://github.com/taskforcesh/bullmq/commit/cbfe437a9aa967bd960890f8ddfb9b01a8c83505))
+
+## [5.72.1](https://github.com/taskforcesh/bullmq/compare/v5.72.0...v5.72.1) (2026-04-02)
+
+
+### Bug Fixes
+
+* **cancellation:** use native AbortController implementation if found ([#3903](https://github.com/taskforcesh/bullmq/issues/3903)) ([46880af](https://github.com/taskforcesh/bullmq/commit/46880af9629e9aee9abbc2782074a3e76d17b924))
+
 # [5.72.0](https://github.com/taskforcesh/bullmq/compare/v5.71.1...v5.72.0) (2026-04-01)